refactor: move post display logic to viewmodel
